Fix RatingControl binding type mismatch
Changed Rating property from int? to double? in both FeedbackSubmission model and FeedbackViewModel to resolve WMC1121 error. RatingControl.Value requires double type for proper binding.

+++ InstallVibe/Models/FeedbackSubmission.cs
@@ -19,7 +19,7 @@ public class FeedbackSubmission
 
     public FeedbackType FeedbackType { get; set; }
 
-    public int? Rating { get; set; } // 1-5 stars
+    public double? Rating { get; set; } // 1-5 stars
 
     public FeedbackCategory Category { get; set; }
